[refactor] move ai code into their own directories
