+++ functions/ai/gemini/geminiPrompts.js
