+++ functions/ai/openai/novel.js
