+++ functions/ai/openai/prompts.js
